Create WikiWorker class
This small powerful change passes the service worker registration
directly to the WikiWorker.

To avoid massive indents and make my code easier to read I add a
getRegisteredWorker function - but there's no need for it to be async :)

A new WikiWorker gets attached to the button and this clears the way for
me not having to work in global scope.

See https://github.com/slightlyoff/ServiceWorker/issues/357 to avoid
confusion.